+++ src/quickpicks/aiModelPicker.ts
@@ -1,8 +1,11 @@
-import type { QuickPickItem } from 'vscode';
+import type { Disposable, QuickInputButton, QuickPickItem } from 'vscode';
 import { QuickPickItemKind, ThemeIcon, window } from 'vscode';
 import type { AIModel } from '../ai/aiProviderService';
 import type { AIModels, AIProviders } from '../constants';
+import { Commands } from '../constants';
 import type { Container } from '../container';
+import { executeCommand } from '../system/command';
+import { getQuickPickIgnoreFocusOut } from '../system/utils';
 
 export interface ModelQuickPickItem extends QuickPickItem {
 	model: AIModel;
@@ -14,8 +17,7 @@ export async function showAIModelPicker(
 ): Promise<ModelQuickPickItem | undefined> {
 	const models = (await (await container.ai)?.getModels()) ?? [];
 
-	type QuickPickSeparator = { label: string; kind: QuickPickItemKind.Separator };
-	const items: (ModelQuickPickItem | QuickPickSeparator)[] = [];
+	const items: ModelQuickPickItem[] = [];
 
 	let lastProvider: AIProviders | undefined;
 	for (const m of models) {
@@ -23,7 +25,7 @@ export async function showAIModelPicker(
 
 		if (lastProvider !== m.provider.id) {
 			lastProvider = m.provider.id;
-			items.push({ label: m.provider.name, kind: QuickPickItemKind.Separator });
+			items.push({ label: m.provider.name, kind: QuickPickItemKind.Separator } as unknown as ModelQuickPickItem);
 		}
 
 		const picked = m.provider.id === current?.provider && m.id === current?.model;
@@ -34,14 +36,48 @@ export async function showAIModelPicker(
 			// description: m.provider.name,
 			model: m,
 			picked: picked,
-		});
+		} satisfies ModelQuickPickItem);
 	}
 
-	const pick = (await window.showQuickPick(items, {
-		title: 'Choose AI Model',
-		placeHolder: 'Select an AI model to use for experimental AI features',
-		matchOnDescription: true,
-	})) as ModelQuickPickItem | undefined;
+	const quickpick = window.createQuickPick<ModelQuickPickItem>();
+	quickpick.ignoreFocusOut = getQuickPickIgnoreFocusOut();
 
-	return pick;
+	const disposables: Disposable[] = [];
+
+	const ResetAIKeyButton: QuickInputButton = {
+		iconPath: new ThemeIcon('clear-all'),
+		tooltip: 'Reset AI Keys...',
+	};
+
+	try {
+		const pick = await new Promise<ModelQuickPickItem | undefined>(resolve => {
+			disposables.push(
+				quickpick.onDidHide(() => resolve(undefined)),
+				quickpick.onDidAccept(() => {
+					if (quickpick.activeItems.length !== 0) {
+						resolve(quickpick.activeItems[0]);
+					}
+				}),
+				quickpick.onDidTriggerButton(e => {
+					if (e === ResetAIKeyButton) {
+						void executeCommand(Commands.ResetAIKey);
+					}
+				}),
+			);
+
+			quickpick.title = 'Choose AI Model';
+			quickpick.placeholder = 'Select an AI model to use for experimental AI features';
+			quickpick.matchOnDescription = true;
+			quickpick.matchOnDetail = true;
+			quickpick.buttons = [ResetAIKeyButton];
+			quickpick.items = items;
+
+			quickpick.show();
+		});
+
+		return pick;
+	} finally {
+		quickpick.dispose();
+		disposables.forEach(d => void d.dispose());
+	}
 }